Add support for Find task and bug fix for some commands

+++ src/main/java/buttercup/parsers/CommandParser.java
@@ -3,11 +3,10 @@ package buttercup.parsers;
 import buttercup.commands.Command;
 import buttercup.exceptions.ButtercupException;
 import buttercup.storage.Storage;
-import buttercup.tasks.Deadline;
-import buttercup.tasks.Event;
-import buttercup.tasks.Task;
-import buttercup.tasks.Todo;
+import buttercup.tasks.*;
 import buttercup.utils.DateTimeFormatUtils;
+
+import java.util.List;
 
 public class CommandParser {
     private final Storage storage;
@@ -24,6 +23,9 @@ public class CommandParser {
             break;
         case MARK:
             try {
+                if (!input.startsWith("mark ")) {
+                    throw new ButtercupException("Invalid mark command, try mark {taskNumber} instead.");
+                }
                 int taskNumber = Integer.parseInt(input.substring(5).trim());
                 result = mark(taskNumber);
             } catch (NumberFormatException e) {
@@ -34,6 +36,9 @@ public class CommandParser {
             break;
         case UNMARK:
             try {
+                if (!input.startsWith("unmark ")) {
+                    throw new ButtercupException("Invalid unmark command, try unmark {taskNumber} instead.");
+                }
                 int taskNumber = Integer.parseInt(input.substring(7).trim());
                 result = unmark(taskNumber);
             } catch (NumberFormatException e) {
@@ -53,6 +58,9 @@ public class CommandParser {
             break;
         case DELETE:
             try {
+                if (!input.startsWith("delete ")) {
+                    throw new ButtercupException("Invalid delete command, try delete {taskNumber} instead.");
+                }
                 int taskNumber = Integer.parseInt(input.substring(7).trim());
                 result = deleteTask(taskNumber);
             } catch (NumberFormatException e) {
@@ -61,6 +69,16 @@ public class CommandParser {
                 System.out.println(e);
             }
             break;
+        case FIND:
+            try {
+                if (!input.startsWith("find ")) {
+                    throw new ButtercupException("Invalid find command, try find {keyword} instead.");
+                }
+                String keyword =  input.substring(5).trim();
+                result = findTask(keyword);
+            } catch (ButtercupException e) {
+                System.out.println(e);
+            }
         default:
             return result;
         }
@@ -187,5 +205,18 @@ public class CommandParser {
         } else {
             throw new ButtercupException("Invalid command, the description of a " + input + " cannot be left empty. Try event {description} /from {start} /to {end} instead.");
         }
+    }
+
+    private String findTask(String keyword) throws ButtercupException {
+        if (keyword.isEmpty()) {
+            throw new ButtercupException("No keyword provided! Try find {keyword} instead.");
+        }
+        TaskList filteredTasks = new TaskList(storage.getTasks().filterByKeyword(keyword));
+        if (filteredTasks.isEmpty()) {
+            return "There are no tasks matching the keyword: " + keyword;
+        }
+        StringBuilder sb = new StringBuilder("Here are the matching tasks in your list:\n");
+        sb.append(filteredTasks);
+        return sb.toString();
     }
 }
